Lock Stats Module for first 3 months from account creation.
